Find by Id is now working. Thymeleaf templates were updated.

+++ src/main/java/application/dao/ClientRepository.java
@@ -1,7 +1,9 @@
 package application.dao;
 
 import application.entity.Client;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ClientRepository extends CrudRepository<Client, Integer> {
+public interface ClientRepository extends CrudRepository<Client, Long> {
+
 }
